Add metadata to a track from a different track
`-s` now accepts another track which can be used as a metadata source.
This metadata source track needs to be separated using "::" from the
track to be downloaded.

For example:
```
$ spotdl -s "nightcore janji heroes::janji heroes"
```
This will download the nightcore version of the track but the original
track would be used for metadata. Similarly, one may also pass Spotify
URIs or YouTube URLs in either of the two tracks.

Closes #500.