Fix performance issue when ffmpeg outputs a lot of lines
Change the python code that records the ffmpeg output to not
unnecessarily construct a new string of the entire output every time
a new line is read from the process. Instead, construct the string only
once at the end, if it is needed.

Before this commit the `out` string was constructed in a loop every time
a new line was read, which requires O(n^2) time where n is the number of
lines in the output. For every loop iteration, the
`"\n".join(out_buffer)` code constructed a new string by iterating over
all the recorded lines so far. This commit reduces the time complexity
to O(n) because the output string is constructed outside of the loop.

Spotdl runs ffmpeg with debug logging level, which can result in a lot of
recorded output. For example, running the `test_download_song` test in
`test_entry_point.py` on my Linux system with ffmpeg 6.1.1 makes ffmpeg
output a total of 13 MiB of messages, consisting of almost 179 000 lines.

This fix reduces the time needed for `test_download_song` from 779 s to
10 s on my machine, which is a 99% performance increase.