Fixing termux installation
After checking again the script in a clean termux environment the script needed some changes on order to work. It always resulted in a compiler error when trying to build the dependencies of the pydatic component. Turn out its a combination of Rust error and the native environment of termux. It needs the Rust package to properly compile it and the binutils package to be have correctly as it should. The setuptools package is not necessary.